requirements.txt
diffusers>=0.30.0
transformers>=4.44.0
accelerate>=0.34.0
peft>=0.12.0
torch>=2.2.0
torchvision>=0.17.0
pyyaml>=6.0.0
Pillow>=10.0.0
tqdm>=4.66.0
pyarrow>=16.0.0
